refactor(@dpc-sdp/ripple-ui-core): :art: refactor primary nav to have sub component folders
