+++ packages/ripple-ui-core/src/components/primary-nav/components/mega-menu/mega-menu.css
@@ -1,0 +1,359 @@
+/* stylelint-disable no-descending-specificity */
+/* stylelint-disable comment-empty-line-before */
+@import '@dpc-sdp/ripple-ui-core/style/breakpoints';
+
+.rpl-primary-nav__mega-menu {
+  width: 100%;
+  max-height: calc(100% - var(--local-nav-bar-height) - var(--rpl-sp-8));
+  overflow-y: auto;
+
+  @media (--rpl-bp-l) {
+    max-height: calc(100% - var(--local-nav-bar-height) - var(--rpl-sp-9));
+  }
+}
+
+.rpl-primary-nav__mega-menu-quick-links {
+  text-align: right;
+  padding-top: var(--rpl-sp-4);
+  padding-right: var(--rpl-sp-4);
+  padding-left: var(--rpl-sp-4);
+
+  li {
+    margin-bottom: var(--rpl-sp-4);
+
+    &:last-of-type {
+      margin-bottom: 0;
+    }
+  }
+
+  a {
+    color: white;
+    text-decoration: none;
+    white-space: nowrap;
+
+    span {
+      white-space: normal;
+    }
+  }
+}
+
+.rpl-primary-nav__mega-menu-user-action {
+  /* Hide the mega menu user action slot when the desktop version is active
+  because it will appear in the nav bar */
+  .rpl-primary-nav--collapse-until-l & {
+    @media (--rpl-bp-l) {
+      display: none;
+    }
+  }
+
+  .rpl-primary-nav--collapse-until-xl & {
+    @media (--rpl-bp-xl) {
+      display: none;
+    }
+  }
+}
+
+.rpl-primary-nav__mega-menu-grid-outer {
+  padding-top: var(--rpl-sp-9);
+  padding-bottom: var(--rpl-sp-8);
+  padding-right: var(--rpl-sp-4);
+  padding-left: var(--rpl-sp-4);
+
+  @media (--rpl-bp-s) {
+    padding-right: var(--rpl-sp-5);
+    padding-left: var(--rpl-sp-5);
+  }
+
+  @media (--rpl-bp-l) {
+    padding-top: var(--rpl-sp-7);
+    padding-right: var(--rpl-sp-10);
+    padding-bottom: var(--rpl-sp-9);
+    padding-left: var(--rpl-sp-10);
+  }
+
+  @media (--rpl-bp-xl) {
+    padding-right: 0;
+    padding-left: 0;
+  }
+}
+
+.rpl-primary-nav__mega-menu-grid-container {
+  width: 100%;
+  max-width: 400px;
+  margin: 0 auto;
+  overflow-x: hidden;
+
+  @media (--rpl-bp-l) {
+    max-width: none;
+  }
+
+  @media (--rpl-bp-xl) {
+    max-width: 1056px;
+  }
+
+  .rpl-primary-nav--collapse-until-xl & {
+    @media (--rpl-bp-l) {
+      max-width: 400px;
+    }
+
+    @media (--rpl-bp-xl) {
+      max-width: 1056px;
+    }
+  }
+
+  .rpl-primary-nav--collapse-always & {
+    @media (--rpl-bp-l) {
+      max-width: 400px;
+    }
+  }
+}
+
+.rpl-primary-nav__mega-menu-grid {
+  display: flex;
+  /* width = 4 full width columns with sp-4 gap between them */
+  width: calc(400%);
+  transition: transform var(--rpl-motion-speed-8) ease;
+
+  .rpl-primary-nav--collapse-until-l & {
+    @media (--rpl-bp-l) {
+      display: grid;
+      width: 100%;
+      transition: none;
+    }
+  }
+
+  .rpl-primary-nav--collapse-until-xl & {
+    @media (--rpl-bp-xl) {
+      display: grid;
+      width: 100%;
+      transition: none;
+    }
+  }
+
+  .rpl-primary-nav__mega-menu--current-level-2 & {
+    /* show second column */
+    transform: translateX(calc(-25%));
+
+    .rpl-primary-nav--collapse-until-l & {
+      @media (--rpl-bp-l) {
+        transform: none;
+      }
+    }
+
+    .rpl-primary-nav--collapse-until-xl & {
+      @media (--rpl-bp-xl) {
+        transform: none;
+      }
+    }
+  }
+
+  .rpl-primary-nav__mega-menu--current-level-3 & {
+    /* show third column */
+    transform: translateX(calc(-50%));
+
+    .rpl-primary-nav--collapse-until-l & {
+      @media (--rpl-bp-l) {
+        transform: none;
+      }
+    }
+
+    .rpl-primary-nav--collapse-until-xl & {
+      @media (--rpl-bp-xl) {
+        transform: none;
+      }
+    }
+  }
+
+  .rpl-primary-nav__mega-menu--current-level-4 & {
+    /* show fourth column */
+    transform: translateX(calc(-75%));
+
+    .rpl-primary-nav--collapse-until-l & {
+      @media (--rpl-bp-l) {
+        transform: none;
+      }
+    }
+
+    .rpl-primary-nav--collapse-until-xl & {
+      @media (--rpl-bp-xl) {
+        transform: none;
+      }
+    }
+  }
+
+  > div {
+    width: 100%;
+  }
+}
+
+.rpl-primary-nav__mega-menu-column {
+  /* The first column goes wider on 'l' breakpoint */
+  &:first-of-type {
+    .rpl-primary-nav--collapse-until-l & {
+      @media (--rpl-bp-l) {
+        grid-column-end: span 12;
+      }
+
+      @media (--rpl-bp-xl) {
+        grid-column-end: span 3;
+      }
+    }
+  }
+
+  .rpl-primary-nav--collapse-until-l & {
+    @media (--rpl-bp-l) {
+      grid-column-end: span 4;
+    }
+
+    @media (--rpl-bp-xl) {
+      grid-column-end: span 3;
+    }
+  }
+
+  .rpl-primary-nav--collapse-until-xl & {
+    @media (--rpl-bp-xl) {
+      grid-column-end: span 3;
+    }
+  }
+}
+
+.rpl-primary-nav__mega-menu-back-button {
+  display: flex;
+  align-items: center;
+  margin-bottom: var(--rpl-sp-8);
+  text-align: left;
+  cursor: pointer;
+
+  .rpl-primary-nav--collapse-until-l & {
+    @media (--rpl-bp-l) {
+      display: none;
+    }
+  }
+
+  .rpl-primary-nav--collapse-until-xl & {
+    @media (--rpl-bp-xl) {
+      display: none;
+    }
+  }
+
+  .rpl-icon {
+    transform: translateY(-1px);
+  }
+}
+
+.rpl-primary-nav__mega-menu-section-title {
+  padding: 0 var(--rpl-sp-2);
+
+  &.rpl-primary-nav__mega-menu-section-title--mobile {
+    margin-bottom: var(--rpl-sp-4);
+
+    .rpl-primary-nav--collapse-until-l & {
+      @media (--rpl-bp-l) {
+        display: none;
+      }
+    }
+
+    .rpl-primary-nav--collapse-until-xl & {
+      @media (--rpl-bp-xl) {
+        display: none;
+      }
+    }
+  }
+
+  &.rpl-primary-nav__mega-menu-section-title--desktop {
+    display: none;
+
+    .rpl-primary-nav--collapse-until-l & {
+      @media (--rpl-bp-l) {
+        display: block;
+        grid-column-end: span 12;
+      }
+    }
+
+    .rpl-primary-nav--collapse-until-xl & {
+      @media (--rpl-bp-xl) {
+        display: block;
+        grid-column-end: span 3;
+      }
+    }
+  }
+}
+
+.rpl-primary-nav__mega-menu-list {
+  /* Hide the level 1 mega menu list when the desktop version is active because
+  the nav bar will show it instead */
+  &.rpl-primary-nav__mega-menu-list--level-1 {
+    .rpl-primary-nav--collapse-until-l & {
+      @media (--rpl-bp-l) {
+        display: none;
+      }
+    }
+
+    .rpl-primary-nav--collapse-until-xl & {
+      @media (--rpl-bp-xl) {
+        display: none;
+      }
+    }
+  }
+
+  .rpl-primary-nav__mega-menu--current-level-1
+    &.rpl-primary-nav__mega-menu-list--level-1 {
+    display: block;
+  }
+
+  .rpl-primary-nav__mega-menu--current-level-2
+    &.rpl-primary-nav__mega-menu-list--level-2 {
+    display: block;
+  }
+
+  .rpl-primary-nav__mega-menu--current-level-3
+    &.rpl-primary-nav__mega-menu-list--level-3 {
+    display: block;
+  }
+
+  .rpl-primary-nav__mega-menu--current-level-4
+    &.rpl-primary-nav__mega-menu-list--level-4 {
+    display: block;
+  }
+}
+
+.rpl-primary-nav__mega-menu-action {
+  display: flex;
+  justify-content: space-between;
+  position: relative;
+  width: 100%;
+  padding: var(--rpl-sp-4) var(--rpl-sp-2);
+  text-align: left;
+  text-decoration: none;
+  color: var(--rpl-clr-type-primary-contrast);
+  border-bottom: 1px solid var(--rpl-clr-primary);
+  cursor: pointer;
+
+  &:hover,
+  &:active,
+  &.rpl-primary-nav__mega-menu-action--active {
+    background-color: var(--rpl-clr-primary-alpha);
+  }
+
+  &.rpl-primary-nav__mega-menu-action--active {
+    &::after {
+      display: block;
+      content: '';
+      position: absolute;
+      bottom: 0;
+      left: 0;
+      border-bottom: 2px solid var(--rpl-clr-type-primary-contrast);
+      width: 100%;
+    }
+  }
+
+  &.rpl-primary-nav__mega-menu-action--home {
+    justify-content: flex-start;
+    align-items: center;
+    font-weight: var(--rpl-type-weight-bold);
+  }
+}
+
+.rpl-primary-nav__mega-menu-action-icon {
+  transform: translateY(2px);
+}
