+++ packages/ripple-ui-core/src/components/primary-nav/components/nav-bar/nav-bar-action.vue
