+++ packages/ripple-ui-core/src/components/primary-nav/components/search-form/search-form.css
@@ -1,0 +1,33 @@
+@import '@dpc-sdp/ripple-ui-core/style/breakpoints';
+
+.rpl-primary-nav__search-form-quick-links {
+  text-align: right;
+  padding-top: var(--rpl-sp-4);
+  padding-right: var(--rpl-sp-4);
+  padding-left: var(--rpl-sp-4);
+}
+
+.rpl-primary-nav__search-bar-wrapper {
+  margin: 0 auto;
+  padding-top: var(--rpl-sp-10);
+  padding-right: var(--rpl-sp-6);
+  padding-bottom: var(--rpl-sp-5);
+  padding-left: var(--rpl-sp-6);
+
+  @media (--rpl-bp-s) {
+    padding-right: var(--rpl-sp-9);
+    padding-left: var(--rpl-sp-9);
+  }
+
+  @media (--rpl-bp-m) {
+    padding-top: var(--rpl-sp-13);
+    padding-right: var(--rpl-sp-11);
+    padding-left: var(--rpl-sp-11);
+  }
+
+  @media (--rpl-bp-l) {
+    max-width: 720px;
+    padding-right: 0;
+    padding-left: 0;
+  }
+}
